# Faster instance creation for ValuesJsonConverter