@Turnerj 2019-12-24
As far as deserialization goes, there are probably two more low-hanging fruit in the library:

1. Optimising the `IEnumerable<object>` path in `Values` to not allocate 4 lists if they aren't all necessary (I've got a few ideas I need to test out with that)

2. Optimising the single-item path for `Values` and `OneOrMany` when coming from `ValuesJsonConverter`. Currently the converter simplifies itself and allocates an array for one item. In `OneOrMany` as an example, it allocates a second array for filtering - that is an OK compromise for a many items but results in one-more-than-necessary allocations for a single item. Really, both types could do with a type `object` constructor - need to see if it is worth it though.