@RehanSaeed 2019-12-30
I discovered the same thing some time ago. I also discovered that the generic `new()` constraint uses `Activator.CreateInstance` under the hood, so generic code can also be speeded up. I ended up using this implementation that the great Jon Skeet actually pointed me at:

https://github.com/Dotnet-Boxed/Framework/blob/master/Source/Boxed.Mapping/Factory.cs

Very similar to yours. I raised an issue in corefx at the time for `new()` generic constraing to be speeded up by avoiding the use of `Activator.CreateInstance` but it hasn't gained much traction.